consolidate login and signup pages #29
