+++ gemini_proact_flutter/lib/view/Auth/auth_page.dart
@@ -1,8 +1,10 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-
+import 'package:logging/logging.dart' show Logger;
 import 'package:gemini_proact_flutter/view/Home/home_page.dart';
 import 'package:gemini_proact_flutter/view/Login-Signup/login_or_signup_page.dart';
+
+final logger = Logger((AuthPage).toString());
 
 class AuthPage extends StatelessWidget {
   const AuthPage({super.key});
@@ -13,11 +15,15 @@ class AuthPage extends StatelessWidget {
       body: StreamBuilder<User?>(
         stream: FirebaseAuth.instance.authStateChanges(),
         builder: (context, snapshot) {
-          if (snapshot.hasData) { // user logged in 
+          if (snapshot.hasData) {
+            // user logged in
+            logger.info('user logged in as name=${snapshot.data?.displayName} email=${snapshot.data?.email}');
             return HomePage(title: 'Proact Home Page');
           }
-          else {                  // user not logged in 
-            return LoginOrSignupPage();
+          else {
+            // user not logged in
+            logger.info('user not logged in');
+            return const LoginOrSignupPage();
           }
         },
       )
